ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
python-multipart==0.0.12
transformers==4.46.3
torch==2.4.1+cpu
accelerate==1.1.1
sentence-transformers==3.2.1
pypdf==5.1.0
python-docx==1.1.2
huggingface-hub==0.26.2
datasets==3.1.0
peft==0.13.2
soundfile==0.12.1
numpy==2.1.3
scipy==1.14.1
orjson==3.10.7
tenacity==9.0.0
opencv-python-headless==4.10.0.84
Pillow==10.4.0
prometheus-client==0.21.1
